% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital_gene_expression.R
\name{digital_gene_expression}
\alias{digital_gene_expression}
\title{Digital Gene Expression.}
\usage{
digital_gene_expression(data)
}
\arguments{
\item{data}{list consisting of the count data and sample table.}
}
\value{
list containing a design matrix and a DGElist object.
}
\description{
Creates a Digital gene expression list and a design matrix.
}
