% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{enrich_analysis}
\alias{enrich_analysis}
\title{Enrich Analysis.}
\usage{
enrich_analysis(
  counts_file,
  sample_table_file,
  db,
  ontology = NULL,
  threshold = 1
)
}
\arguments{
\item{counts_file}{file containing count data.}

\item{db}{"KEGG" or "GO"}

\item{ontology}{"BP", "CC" or "MF"}

\item{threshold}{minimum mean-log2-CPM to keep the data during low expression filtering.}

\item{sample_table}{file containing sample table.}
}
\description{
Calls upon all functions of the package to instantly perform EA and save the result plots to the users working directory.
}
