% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_Hs_genes.R
\name{enrich_Hs_genes}
\alias{enrich_Hs_genes}
\title{Enrich genes with KEGG or GO pathways (Homo sapiens only)}
\usage{
enrich_Hs_genes(entrez, db, ontology = NULL)
}
\arguments{
\item{entrez}{vector with ENTREZID's to enrich}

\item{db}{"KEGG" or "GO"}

\item{ontology}{"BP", "CC" or "MF"}
}
\value{
enrichResult class object
}
\description{
Returns a enrichResult class using a vector of ENTREZID's and specifying
which database to enrich from (KEGG or GO) using enrichGO or enrichKEGG
function from clusterProfiler. If db = "GO", ontology must be chosen.
Biological Processes ("BP"), Cellular Components ("CC") or Molecular
Functions ("MF").
}
