% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_plots.R
\name{enrich_plots}
\alias{enrich_plots}
\title{Plots dot-, bar-, cnet-, upset-, go- and treeplot for
over-representaiton data.}
\usage{
enrich_plots(rich_res, n, font_size = NULL, file_id, ...)
}
\arguments{
\item{rich_res}{enrichResult class}

\item{n}{number of annotations}

\item{font_size}{font size}

\item{file_id}{unique id for plots}

\item{...}{width =, height =, dpi =,}
}
\value{
multiple plots.png
}
\description{
Each plot will be saved to a .png following this naming convention:
(file id)_(database type)_(plot type).png.
Database type and plot type is set automatically, user must set unique id.
User also has the ability to change the font size and number of annotations.
Furthermore, user can decide width, depth and dpi of .png file by passing the
dots. The plots are saved in the current working directory.
}
