% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data_local.R
\name{fetch_data_from_file}
\alias{fetch_data_from_file}
\title{Title Fetch data from file}
\usage{
fetch_data_from_file(counts_file, sample_table_file)
}
\arguments{
\item{sample_table}{}
}
\description{
Title Fetch data from file
}
