% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_test_model.R
\name{fit_test_model}
\alias{fit_test_model}
\title{Linear regression using generalized linear model}
\usage{
fit_test_model(digital_deg_list)
}
\arguments{
\item{digital_deg_list}{list containing a design matrix and a DGElist object.}
}
\value{
Digital Gene Expression Likelihood Ratio Test data and results
- class (DGELRT)
}
\description{
Estimates common, trended and tagwise dispersions and fits a general
linear model to test for differentially expressed genes using functions from
the edgeR and limma package
}
