% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deg.R
\name{get_deg}
\alias{get_deg}
\title{Aquires the differently expressed genes (DEGs).}
\usage{
get_deg(deg_results, lfc = 1, adj_p = 0.05)
}
\arguments{
\item{deg_results}{DGELRT object}

\item{lfc}{numeric value indicating min log fold change. Default = 1}

\item{adj_p}{numeric value indicating max adjusted p value. Default 0.05}
}
\value{
data.frame
}
\description{
Returns a data frame with DEGS meeting the inputed log fold change and
adjusted p-values (FDR) conditions.
}
