% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Hs_entrez.R
\name{get_Hs_entrez}
\alias{get_Hs_entrez}
\title{Adds ENTREZID column to data frame}
\usage{
get_Hs_entrez(df, key = "SYMBOL")
}
\arguments{
\item{df}{data frame}

\item{key}{keytype, default = "SYMBOL"}
}
\value{
dataframe$ENTREZID
}
\description{
Will map a column with ENTREZID for each gene based on the rownames keytype
(ENTREZID, SYMBOL, ENSEMBL) of data frame.
}
