% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_deg_excel.R
\name{save_deg_excel}
\alias{save_deg_excel}
\title{Save the Differentially Expressed Genes data.frame to excel.}
\usage{
save_deg_excel(deg_df, filename)
}
\arguments{
\item{deg_df}{data.frame with differently expressed genes}

\item{filename}{string naming the excel file. (.xlsx is added automatically)}
}
\value{
filename.xlsx
}
\description{
Creates a stylish excel file based on the DEG data.frame using openxlsx. The
file extension (.xslx) is automatically added to inputed filename.
}
