% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data_local.R
\name{fetch_data_from_file}
\alias{fetch_data_from_file}
\title{Fetch data from file.}
\usage{
fetch_data_from_file(counts_file, sample_table_file)
}
\arguments{
\item{counts_file}{file containing count data.}
\item{sample_table}{file containing sample table.}
}
\value{
list with a count data-frame and a sample table data-frame.
}
\description{
Imports transcriptomics count data from local text files.
Expects two files, a count data file and a sample table file.
}
