% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_test_model.R
\name{fit_test_model}
\alias{fit_test_model}
\title{Linear regression using generalized linear model}
\usage{
fit_test_model(dge)
}
\arguments{
\item{dge_list}{DGEList object}

\item{design_matrix}{numeric design matrix}
}
\value{
Digital Gene Expression Likelihood Ratio Test data and results
- class (DGELRT)
}
\description{
Estimates common, trended and tagwise dispersions and fits a general
linear model to test for differentially expressed genes using functions from
the edgeR and limma package
}
