% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Hs_entrez.R
\name{get_Hs_entrez}
\alias{get_Hs_entrez}
\title{Adds ENTREZID column to data frame}
\usage{
get_Hs_entrez(df, key = "SYMBOL")
}
\arguments{
\item{df}{data frame with SYMBOL or ENSEMBL rownames.}

\item{key}{keytype, default = "SYMBOL"}
}
\value{
dataframe with appended ENTREZID column
}
\description{
Will map a column with ENTREZID for each gene based on the rownames keytype
(SYMBOL, ENSEMBL) of data frame. ketype is defaulted to SYMBOL.
}
